Move test-only init_db out of production database.py (ENG-709)
